requirements.txt
flask
requests
ollama
sentence-transformers
faiss-cpu
pypdf